+++ device-bootloaders/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- [SW-15175] Removing mem_size from flash config data and reading ddr size and vendor ID correctly.
 ### Deprecated
 ### Removed
 ### Fixed